Moved more code from TardisMod Split darkcraft and darkcore

+++ src/main/java/io/darkcraft/darkcore/mod/abstracts/AbstractBlockRenderer.java
@@ -31,7 +31,7 @@ public abstract class AbstractBlockRenderer extends TileEntitySpecialRenderer
 		int l1 = l % 65536;
 		int l2 = l / 65536;
 		tessellator.setColorOpaque_F(brightness, brightness, brightness);
-		OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, (float)l1, (float)l2);
+		OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, l1, l2);
 		/*Note that true tile entity coordinates (tileEntity.xCoord, etc) do not match to render coordinates (d, etc) that are calculated as [true coordinates] - [player coordinates (camera coordinates)]*/
 		renderBlock(tessellator,tileEntity,x,y,z);
 		GL11.glPopMatrix();
